+++ Packages/Tracking Preview/Locomotion/Runtime/Scripts/Teleportation/Teleport Action/Jump Gem Teleport/JumpGem.cs
@@ -171,11 +171,13 @@ namespace Leap.Unity.Preview.Locomotion
 
             if (leapProvider == null)
             {
-                leapProvider = FindAnyObjectByType<LeapProvider>();
+                leapProvider = Hands.Provider;
             }
-
+#if UNITY_2021_3_18_OR_NEWER
             _jumpGemTeleport = FindAnyObjectByType<JumpGemTeleport>(FindObjectsInactive.Include);
-
+#else
+            _jumpGemTeleport = FindObjectOfType<JumpGemTeleport>(true);
+#endif
             if (_audioSource == null)
             {
                 _audioSource = GetComponentInChildren<AudioSource>(true);