+++ Packages/Tracking Preview/Locomotion/Runtime/Scripts/Teleportation/Teleport Action/Jump Gem Teleport/JumpGemTeleport.cs
@@ -48,7 +48,11 @@ namespace Leap.Unity.Preview.Locomotion
 
         private void Awake()
         {
+#if UNITY_2021_3_18_OR_NEWER
             _jumpGems = FindObjectsByType<JumpGem>(FindObjectsInactive.Include, FindObjectsSortMode.None).ToList();
+#else
+            _jumpGems = FindObjectsOfType<JumpGem>(true).ToList();
+#endif
             for (int i = 0; i < _jumpGems.Count; i++)
             {
                 int j = i;