+++ Packages/Tracking/Core/Runtime/Plugins/LeapCSharp/LeapC.cs
@@ -1122,12 +1122,15 @@ namespace LeapInternal
         [DllImport("LeapC", EntryPoint = "LeapDestroyConnection")]
         public static extern void DestroyConnection(IntPtr connection);
 
+        [Obsolete("Config is not used in Ultraleap's Tracking Service 5.X+. This will be removed in the next Major release")]
         [DllImport("LeapC", EntryPoint = "LeapSaveConfigValue")]
         private static extern eLeapRS SaveConfigValue(IntPtr hConnection, string key, IntPtr value, out UInt32 requestId);
 
+        [Obsolete("Config is not used in Ultraleap's Tracking Service 5.X+. This will be removed in the next Major release")]
         [DllImport("LeapC", EntryPoint = "LeapRequestConfigValue")]
         public static extern eLeapRS RequestConfigValue(IntPtr hConnection, string name, out UInt32 request_id);
 
+        [Obsolete("Config is not used in Ultraleap's Tracking Service 5.X+. This will be removed in the next Major release")]
         public static eLeapRS SaveConfigValue(IntPtr hConnection, string key, bool value, out UInt32 requestId)
         {
             LEAP_VARIANT_VALUE_TYPE valueStruct = new LEAP_VARIANT_VALUE_TYPE(); //This is a C# approximation of a C union
@@ -1135,6 +1138,7 @@ namespace LeapInternal
             valueStruct.boolValue = value ? 1 : 0;
             return SaveConfigWithValueType(hConnection, key, valueStruct, out requestId);
         }
+        [Obsolete("Config is not used in Ultraleap's Tracking Service 5.X+. This will be removed in the next Major release")]
         public static eLeapRS SaveConfigValue(IntPtr hConnection, string key, Int32 value, out UInt32 requestId)
         {
             LEAP_VARIANT_VALUE_TYPE valueStruct = new LEAP_VARIANT_VALUE_TYPE();
@@ -1142,6 +1146,7 @@ namespace LeapInternal
             valueStruct.intValue = value;
             return SaveConfigWithValueType(hConnection, key, valueStruct, out requestId);
         }
+        [Obsolete("Config is not used in Ultraleap's Tracking Service 5.X+. This will be removed in the next Major release")]
         public static eLeapRS SaveConfigValue(IntPtr hConnection, string key, float value, out UInt32 requestId)
         {
             LEAP_VARIANT_VALUE_TYPE valueStruct = new LEAP_VARIANT_VALUE_TYPE();
@@ -1149,6 +1154,7 @@ namespace LeapInternal
             valueStruct.floatValue = value;
             return SaveConfigWithValueType(hConnection, key, valueStruct, out requestId);
         }
+        [Obsolete("Config is not used in Ultraleap's Tracking Service 5.X+. This will be removed in the next Major release")]
         public static eLeapRS SaveConfigValue(IntPtr hConnection, string key, string value, out UInt32 requestId)
         {
             LEAP_VARIANT_REF_TYPE valueStruct;
@@ -1156,6 +1162,8 @@ namespace LeapInternal
             valueStruct.stringValue = value;
             return SaveConfigWithRefType(hConnection, key, valueStruct, out requestId);
         }
+
+        [Obsolete("Config is not used in Ultraleap's Tracking Service 5.X+. This will be removed in the next Major release")]
         private static eLeapRS SaveConfigWithValueType(IntPtr hConnection, string key, LEAP_VARIANT_VALUE_TYPE valueStruct, out UInt32 requestId)
         {
             IntPtr configValue = Marshal.AllocHGlobal(Marshal.SizeOf(valueStruct));
@@ -1171,6 +1179,8 @@ namespace LeapInternal
             }
             return callResult;
         }
+
+        [Obsolete("Config is not used in Ultraleap's Tracking Service 5.X+. This will be removed in the next Major release")]
         private static eLeapRS SaveConfigWithRefType(IntPtr hConnection, string key, LEAP_VARIANT_REF_TYPE valueStruct, out UInt32 requestId)
         {
             IntPtr configValue = Marshal.AllocHGlobal(Marshal.SizeOf(valueStruct));