+++ Packages/Tracking/CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Example content to be split by XR, Tabletop and URP Examples
 - Combined example content to be part of the main Ultraleap Tracking .unitypackage when importing via .unitypackage
 - Added fog and gradient sky for all XR example scenes
+- (Hand Rays) Exposed dot product used to test if the hand is facing camera
 
 ### Fixed
 - Errors in Editor when using pre-2023.3.18 LTS due to FindObjectByType issue